process execute changes
now process execute passes a pointer to a struct to thread_create instead of a pointer to a string, the struct contains a semaphore used by the parent to ensure that it waits untill its child has finished loading and return -1 if it failed to load. start_process now, on failure, sets the loaded boolean to false if not successful, and will always call sema_up to wake up the parent thread.